Extend coverage of padding and bug fix in 1d case
When passing in the desired number of padding elements in the 1D case,
it worked well if given an integer. But given a length 1 tuple, e.g.
(10,), it failed because the code added the tuple to the final shape,
not the integer value on it. The fix was a case of changing a ``append``
by an ``extend``. New tests caught this bug and now have it covered.